hr/requirements.txt
APScheduler
arabic-reshaper
asgiref
asn1crypto
bs4
certifi
cffi
chardet<6
charset-normalizer
click
cryptography
cssselect2
Django==5.2
django-auditlog
django-apscheduler
django-cors-headers
django-environ
django-filter
django-import-export
django-jsonfield
django-mathfilters
django-microsoft-auth
django-model-utils
django-simple-history
django-widget-tweaks
djangorestframework
djangorestframework-simplejwt
drf-yasg
et-xmlfile
geopy
google-api-python-client
google-auth-oauthlib
google-cloud-storage==3.0.0
html5lib
Jinja2
idna
lxml
numpy
oauthlib
openpyxl
oscrypto
pandas
pdfkit
pillow
pycparser
pyHanko
pyhanko-certvalidator
PyMuPDF
pypdf
pypng
python-bidi
python-dateutil
pytz
PyYAML
pyzk==0.9
qrcode
reportlab==4.0.9
requests
responses
setuptools
six
sqlparse
svglib
swapper
tinycss2
typing_extensions
tzdata
tzlocal
uritools
urllib3
utils
webencodings
xhtml2pdf
XlsxWriter
gunicorn
psycopg2-binary
whitenoise
django-storages
